% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{msu_palettes}
\alias{msu_palettes}
\alias{msu_palettes_discrete}
\alias{msu_palettes_sequential}
\title{MSU Palettes}
\format{
## `msu_palettes`
An object of class `palettes_palette` with `r length(msu_palettes)`
colour palettes. Use `names(msu_palettes)` to return all palette names.

## `msu_palettes_discrete`
An object of class `palettes_palette` with
`r length(msu_palettes_discrete)` discrete colour palettes.
 Use `names(msu_palettes_discrete)` to return all palette names.

## `msu_palettes_sequential`
An object of class `palettes_palette` with
`r length(msu_palettes_sequential)` sequential colour palettes.
 Use `names(msu_palettes_sequential)` to return all palette names.
}
\usage{
msu_palettes

msu_palettes_discrete

msu_palettes_sequential
}
\description{
Palettes from MSU branding colors and other useful colors.
}
\examples{
# Get all palettes by name.
names(msu_palettes)

# Plot all palettes.
plot(msu_palettes)
}
\keyword{datasets}
