% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msuihp-package.R
\docType{package}
\name{msuihp-package}
\alias{msuihp}
\alias{msuihp-package}
\title{msuihp: Helper functions and color palettes for Michigan State University Institute for Health Policy}
\description{
Includes helper functions for common plotting. Provides several color palettes that use MSU branding colors for easy access.
}
\keyword{internal}
